requirements.txt
base58check==1.0.2
ecdsa==0.16.1
htmlmin==0.1.12
python-dotenv==0.15.0
requests==2.22.0
